Add new models in InspectorGadget application
This commit adds primary models to the InspectorGadget application, including its relations as well as enums. Additions include models for device repairs, statuses, employees, requests and client data, and aim to construct a database context for the connection. These models will now lay the foundation to carry out CRUD operations and build the primary business logic of the application. It also includes the addition of a .gitignore file for the project to curb versioning of unnecessary files.